Move client to base class
